+++ src/main/java/insane96mcp/enhancedai/modules/mobs/anticheese/BreakVehicleGoal.java
@@ -32,7 +32,7 @@ public class BreakVehicleGoal extends Goal {
 	}
 
 	private boolean hasValidVehicle() {
-		return this.mob.getVehicle() != null && this.mob.getVehicle().getType().is(AntiCheese.ANTI_CHEESE_VEHICLES);
+		return this.mob.getVehicle() != null && this.mob.getVehicle().getType().is(AntiCheese.VALID_VEHICLES);
 	}
 
 	public void start() {