+++ src/main/java/insane96mcp/enhancedai/setup/NBTUtils.java
@@ -9,6 +9,7 @@ public class NBTUtils {
     /**
      * Returns the int read from the compoundTag or if absent puts the defaultValue in the compoundTag and returns it
      */
+    @Deprecated
     public static int getIntOrPutDefaultLegacy(CompoundTag compoundTag, String tagName, int defaultValue) {
         int result = defaultValue;
         if (compoundTag.contains(tagName)) {
@@ -23,6 +24,7 @@ public class NBTUtils {
     /**
      * Returns the double read from the compoundTag or if absent puts the defaultValue in the compoundTag and returns it
      */
+    @Deprecated
     public static double getDoubleOrPutDefaultLegacy(CompoundTag compoundTag, String tagName, double defaultValue) {
         double result = defaultValue;
         if (compoundTag.contains(tagName)) {
@@ -37,6 +39,7 @@ public class NBTUtils {
     /**
      * Returns the boolean read from the compoundTag or if absent puts the defaultValue in the compoundTag and returns it
      */
+    @Deprecated
     public static boolean getBooleanOrPutDefaultLegacy(CompoundTag compoundTag, String tagName, boolean defaultValue) {
         boolean result = defaultValue;
         if (compoundTag.contains(tagName)) {
@@ -49,7 +52,7 @@ public class NBTUtils {
     }
 
     /**
-     * Returns the boolean read from the compoundTag or if absent puts the defaultValue in the compoundTag and returns it
+     * Returns the boolean read from the compoundTag or if absent, puts the defaultValue in the compoundTag and returns it
      */
     public static boolean getBooleanOrPutDefault(Entity entity, ResourceLocation key, boolean defaultValue) {
         if (!ModNBTData.contains(entity, key)) {
@@ -60,7 +63,7 @@ public class NBTUtils {
     }
 
     /**
-     * Returns the boolean read from the compoundTag or if absent puts the defaultValue in the compoundTag and returns it
+     * Returns the boolean read from the compoundTag or if absent, puts the defaultValue in the compoundTag and returns it
      */
     public static int getIntOrPutDefault(Entity entity, ResourceLocation key, int defaultValue) {
         if (!ModNBTData.contains(entity, key)) {
@@ -68,5 +71,16 @@ public class NBTUtils {
             return defaultValue;
         }
         return ModNBTData.get(entity, key, Integer.class);
+    }
+
+    /**
+     * Returns the boolean read from the compoundTag or if absent, puts the defaultValue in the compoundTag and returns it
+     */
+    public static double getDoubleOrPutDefault(Entity entity, ResourceLocation key, double defaultValue) {
+        if (!ModNBTData.contains(entity, key)) {
+            ModNBTData.put(entity, key, defaultValue);
+            return defaultValue;
+        }
+        return ModNBTData.get(entity, key, Double.class);
     }
 }